arithmetic between different enum types deprecated: convert to int before arithmetics

+++ PWGDQ/reducedTree/AliReducedVarManager.h
@@ -366,8 +366,8 @@ class AliReducedVarManager : public TObject {
     kVZEROTPCoutDiff,
     kVZEROACTotalMult,
     kCorrectedMultiplicity,
-    kNMultiplicityEstimators = kCorrectedMultiplicity - kMultiplicity,
-    kSPDFiredChips = kCorrectedMultiplicity + kNMultiplicityEstimators * ( 1 + kNCorrections * kNReferenceMultiplicities * kNSmearingMethods),  // SPD fired chips in first and second layer
+    kNMultiplicityEstimators = (int)kCorrectedMultiplicity - (int)kMultiplicity,
+    kSPDFiredChips = (int)kCorrectedMultiplicity + (int)kNMultiplicityEstimators * (1 + (int)kNCorrections * (int)kNReferenceMultiplicities * (int)kNSmearingMethods),  // SPD fired chips in first and second layer
     kITSnClusters = kSPDFiredChips+2,       // number of ITS clusters in each layer
     kSPDnSingleClusters = kITSnClusters+6,  // number of clusters in SPD layer 1 not mached to tracklets from layer 2
     kSDDandSSDclusters,                     // number of clusters in the SDD and SSD layers
